requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.1.0
gunicorn==21.2.0
browser-use==0.1.4
torch==2.0.0+cpu --index-url https://download.pytorch.org/whl/cpu
transformers==4.38.0
playwright>=1.40.0
